refactor: remove code duplicaton on board's and user's documents pages
