+++ app/(loggedin)/home/documents/page.tsx
@@ -5,9 +5,11 @@ import { useEffect, useState } from 'react';
 
 import { getUser } from '@/redux/user/userThunk';
 import { useAppDispatch, useAppSelector } from '@/redux/hooks';
-import { TITLE_MAX_LENGTH, FIXED_GRID_STYLES } from '@/data/constants';
-import DocumentCard from '@/components/HomePage/Kanban/Column/JobPosts/JobModal/JobDocuments/DocumentCard';
+import { TITLE_MAX_LENGTH } from '@/data/constants';
+import useDocumentActions from '@/hooks/useDocumentActions';
+import DocumentGrid from '@/components/Documents/DocumentGrid';
 import EditDocumentModal from '@/components/Forms/AddDocument/EditDocumentModal';
+import DocumentFilterBar, { CategoryCount } from '@/components/Documents/DocumentFilterBar';
 import {
   deleteDocument,
   getDocumentsPerUser,
@@ -36,10 +38,7 @@ const UserDocuments = () => {
     profilePicUrl?: string;
   } | null>(null);
   const [selectedCategory, setSelectedCategory] = useState<string | null>(null); // null = All
-  const [isEditModalOpen, setIsEditModalOpen] = useState(false);
-  const [documentToEdit, setDocumentToEdit] = useState<JobDocument | null>(
-    null
-  );
+  // We'll use the document action states from the hook instead of defining them here
 
   // Function to refresh user documents
   const handleDocumentsRefresh = async () => {
@@ -52,6 +51,17 @@ const UserDocuments = () => {
       }
     }
   };
+  
+  // Use the shared document actions hook
+  const {
+    isEditModalOpen,
+    documentToEdit,
+    handleEditDocument,
+    handleDeleteDocument,
+    handleDownloadDocument,
+    setIsEditModalOpen,
+    setDocumentToEdit
+  } = useDocumentActions(userDocuments, accessToken, handleDocumentsRefresh);
 
   // Fetch user info for uploader details
   useEffect(() => {
@@ -136,83 +146,11 @@ const UserDocuments = () => {
       )
     : enhancedDocuments;
 
-  const handleEditDocument = (document: JobDocument) => {
-    // Find the original document with the full title from userDocuments
-    const originalDocument = userDocuments.find(doc => doc.id === document.id);
-    // Use the original document if found, otherwise use the provided document
-    setDocumentToEdit(originalDocument || document);
-    setIsEditModalOpen(true);
-  };
+  // Document action handlers are now provided by the useDocumentActions hook
 
-  const handleDeleteDocument = async (documentId: string) => {
-    try {
-      if (!accessToken) {
-        toast.error('Authentication required');
-        return;
-      }
+  // handleDeleteDocument is now provided by the useDocumentActions hook
 
-      // For user documents, always delete the document entirely from the database
-      // This will automatically detach it from all job applications and remove it completely
-      await dispatch(
-        deleteDocument({
-          documentId,
-          accessToken: accessToken as string,
-        })
-      ).unwrap();
-
-      toast.success('Document deleted successfully!');
-      await handleDocumentsRefresh();
-    } catch (error) {
-      console.error('Error deleting document:', error);
-      toast.error('Failed to delete document. Please try again.');
-    }
-  };
-
-  const handleDownloadDocument = async (jobDocument: JobDocument) => {
-    try {
-      if (!jobDocument.url) {
-        toast.error('Document URL not available');
-        return;
-      }
-
-      const newTab = window.open(
-        jobDocument.url,
-        '_blank',
-        'noopener,noreferrer'
-      );
-
-      if (!newTab || newTab.closed || typeof newTab.closed === 'undefined') {
-        // Fallback: Create a download link for popup blocker case
-        const link = document.createElement('a');
-        link.href = jobDocument.url;
-        link.download = jobDocument.title || 'document';
-        link.target = '_blank';
-        document.body.appendChild(link);
-        link.click();
-        document.body.removeChild(link);
-        toast.success('Document download initiated');
-        return;
-      }
-
-      // Use a more robust approach with proper error handling and longer timeout
-      const checkTabStatus = setTimeout(() => {
-        try {
-          if (newTab && !newTab.closed) {
-            toast.success('Document opened in new tab');
-          } else {
-            toast.success('Document has been downloaded');
-          }
-        } catch (error) {
-          // Tab may be closed, cross-origin, or inaccessible
-          console.warn('Could not check tab status:', error);
-          toast.success('Document has been processed');
-        }
-      }, 1000);
-    } catch (error) {
-      console.error('Error opening document:', error);
-      toast.error('Failed to open document. Please try again.');
-    }
-  };
+  // handleDownloadDocument is now provided by the useDocumentActions hook
 
   if (userDocumentsStatus === 'loading') {
     return (
@@ -230,85 +168,13 @@ const UserDocuments = () => {
         </div>
       </div>
 
-      {/* Filter Bar - above the cards, full width, white bg */}
-      <div className="w-full bg-white pt-6 pb-2 px-6 border-b">
-        <div className="flex items-center gap-2">
-          {/* All filter */}
-          <div
-            className={`flex items-center gap-1 px-3 py-1 rounded-lg cursor-pointer font-medium text-sm transition ${
-              selectedCategory === null
-                ? 'bg-violet-100 text-violet-700 ring-2 ring-violet-300'
-                : 'bg-violet-50 text-violet-700 hover:bg-violet-100'
-            }`}
-            style={{ minWidth: 48 }}
-            onClick={() => setSelectedCategory(null)}
-            tabIndex={0}
-            role="button"
-            aria-pressed={selectedCategory === null}
-          >
-            <span>All</span>
-            <span className="ml-1 px-2 py-0.5 rounded bg-gray-200 text-gray-700 font-semibold text-xs">
-              {userDocuments.length}
-            </span>
-          </div>
-          {/* Category filters */}
-          {categoryCounts.map(({ category, count }) => {
-            // Color mapping for categories
-            const colorMap: Record<string, string> = {
-              Resume: 'bg-blue-100 text-blue-700 hover:bg-blue-200',
-              'Cover Letter': 'bg-green-100 text-green-700 hover:bg-green-200',
-              'Writing Sample':
-                'bg-orange-100 text-orange-700 hover:bg-orange-200',
-              Portfolio: 'bg-purple-100 text-purple-700 hover:bg-purple-200',
-              Recommendation: 'bg-pink-100 text-pink-700 hover:bg-pink-200',
-              'Job Post': 'bg-lime-100 text-lime-700 hover:bg-lime-200',
-              'Offer Letter': 'bg-amber-300 text-amber-800 hover:bg-amber-400',
-              Certification: 'bg-teal-100 text-teal-700 hover:bg-teal-200',
-              Other: 'bg-indigo-100 text-indigo-700 hover:bg-indigo-200',
-              Transcript: 'bg-red-100 text-red-700 hover:bg-red-200',
-              Uncategorized: 'bg-gray-100 text-gray-700 hover:bg-gray-200',
-            };
-            const selectedColorMap: Record<string, string> = {
-              Resume: 'bg-blue-100 text-blue-700 ring-2 ring-blue-300',
-              'Cover Letter':
-                'bg-green-100 text-green-700 ring-2 ring-green-300',
-              'Writing Sample':
-                'bg-orange-100 text-orange-700 ring-2 ring-orange-300',
-              Portfolio: 'bg-purple-100 text-purple-700 ring-2 ring-purple-300',
-              Recommendation: 'bg-pink-100 text-pink-700 ring-2 ring-pink-300',
-              'Job Post': 'bg-lime-100 text-lime-700 ring-2 ring-lime-300',
-              'Offer Letter':
-                'bg-amber-300 text-amber-800 ring-2 ring-amber-500',
-              Certification: 'bg-teal-100 text-teal-700 ring-2 ring-teal-300',
-              Other: 'bg-indigo-100 text-indigo-700 ring-2 ring-indigo-300',
-              Transcript: 'bg-red-100 text-red-700 ring-2 ring-red-300',
-              Uncategorized: 'bg-gray-100 text-gray-700 ring-2 ring-gray-300',
-            };
-            const colorClass =
-              selectedCategory === category
-                ? selectedColorMap[category] ||
-                  'bg-gray-100 text-gray-700 ring-2 ring-gray-300'
-                : colorMap[category] ||
-                  'bg-gray-100 text-gray-700 hover:bg-gray-200';
-            return (
-              <div
-                tabIndex={0}
-                role="button"
-                key={category}
-                style={{ minWidth: 48 }}
-                aria-pressed={selectedCategory === category}
-                onClick={() => setSelectedCategory(category)}
-                className={`flex items-center gap-1 px-3 py-1 rounded-lg cursor-pointer font-medium text-sm transition ${colorClass}`}
-              >
-                <span className="px-2 py-0.5 rounded bg-gray-200 text-gray-700 font-semibold text-xs">
-                  {count}
-                </span>
-                <span>{category}</span>
-              </div>
-            );
-          })}
-        </div>
-      </div>
+      {/* Filter Bar using the shared component */}
+      <DocumentFilterBar
+        allCount={userDocuments.length}
+        categoryCounts={categoryCounts}
+        selectedCategory={selectedCategory}
+        setSelectedCategory={setSelectedCategory}
+      />
 
       {userDocuments.length === 0 ? (
         <div className="flex-1 flex items-center justify-center">
@@ -316,26 +182,14 @@ const UserDocuments = () => {
             You have not created any documents yet
           </p>
         </div>
-      ) : filteredDocuments.length === 0 ? (
-        <div className="flex-1 flex items-center justify-center">
-          <p className="text-center text-xl text-slate-400">
-            No documents found for this category
-          </p>
-        </div>
       ) : (
-        <div className="flex-1 overflow-y-auto p-6">
-          <div style={FIXED_GRID_STYLES}>
-            {filteredDocuments.map((document) => (
-              <DocumentCard
-                key={document.id}
-                document={document}
-                onEdit={handleEditDocument}
-                onDelete={handleDeleteDocument}
-                onDownload={handleDownloadDocument}
-              />
-            ))}
-          </div>
-        </div>
+        <DocumentGrid
+          documents={filteredDocuments}
+          onEdit={handleEditDocument}
+          onDelete={handleDeleteDocument}
+          onDownload={handleDownloadDocument}
+          emptyMessage="No documents found for this category"
+        />
       )}
 
       {/* Edit Document Modal */}
@@ -349,14 +203,7 @@ const UserDocuments = () => {
             setDocumentToEdit(null);
             
             // Then refresh the documents
-            if (accessToken) {
-              try {
-                // Directly dispatch the action to ensure it updates the Redux store
-                await dispatch(getDocumentsPerUser(accessToken));
-              } catch (error) {
-                console.error('Failed to refresh documents after edit:', error);
-              }
-            }
+            await handleDocumentsRefresh();
           }}
           onClose={() => {
             setIsEditModalOpen(false);
