+++ utils/documentColorMaps.ts
@@ -1,0 +1,29 @@
+// Document category color mappings for filter buttons
+export const documentColorMap: Record<string, string> = {
+  Resume: 'bg-blue-100 text-blue-700 hover:bg-blue-200',
+  'Cover Letter': 'bg-green-100 text-green-700 hover:bg-green-200',
+  'Writing Sample': 'bg-orange-100 text-orange-700 hover:bg-orange-200',
+  Portfolio: 'bg-purple-100 text-purple-700 hover:bg-purple-200',
+  Recommendation: 'bg-pink-100 text-pink-700 hover:bg-pink-200',
+  'Job Post': 'bg-lime-100 text-lime-700 hover:bg-lime-200',
+  'Offer Letter': 'bg-amber-300 text-amber-800 hover:bg-amber-400',
+  Certification: 'bg-teal-100 text-teal-700 hover:bg-teal-200',
+  Other: 'bg-indigo-100 text-indigo-700 hover:bg-indigo-200',
+  Transcript: 'bg-red-100 text-red-700 hover:bg-red-200',
+  Uncategorized: 'bg-gray-100 text-gray-700 hover:bg-gray-200',
+};
+
+// Selected state color mappings for filter buttons
+export const selectedDocumentColorMap: Record<string, string> = {
+  Resume: 'bg-blue-100 text-blue-700 ring-2 ring-blue-300',
+  'Cover Letter': 'bg-green-100 text-green-700 ring-2 ring-green-300',
+  'Writing Sample': 'bg-orange-100 text-orange-700 ring-2 ring-orange-300',
+  Portfolio: 'bg-purple-100 text-purple-700 ring-2 ring-purple-300',
+  Recommendation: 'bg-pink-100 text-pink-700 ring-2 ring-pink-300',
+  'Job Post': 'bg-lime-100 text-lime-700 ring-2 ring-lime-300',
+  'Offer Letter': 'bg-amber-300 text-amber-800 ring-2 ring-amber-500',
+  Certification: 'bg-teal-100 text-teal-700 ring-2 ring-teal-300',
+  Other: 'bg-indigo-100 text-indigo-700 ring-2 ring-indigo-300',
+  Transcript: 'bg-red-100 text-red-700 ring-2 ring-red-300',
+  Uncategorized: 'bg-gray-100 text-gray-700 ring-2 ring-gray-300',
+};
